requirements.txt
fastapi==0.103.1
pyngrok==7.0.0
pydantic==1.10.11
langchain==0.0.301
openai==0.28.0
tiktoken==0.5.1
python-telegram-bot==20.5
python-dotenv=1.0.0
google-api-python-client==2.101.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-search-results==2.4.2